requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosql==7.2
alembic==1.9.2
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
bcrypt==4.0.1
billiard==3.6.4.0
boto3==1.26.67
botocore==1.29.67
celery==5.2.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
kombu==5.2.4
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
passlib==1.7.4
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.2
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2